authentication API restructured and modified test project
