+++ server.js
@@ -5,19 +5,10 @@ var app = express();                 // define our app using express
 var bodyParser = require('body-parser');
 var userData = require('./user');
 var fs = require("fs");
-
 var server = require('http').createServer(app);
 var io = require('socket.io')(server);
-var jwt = require('jsonwebtoken');
-
 var sockets = require('./socket/socket');
 app.use(express.static(__dirname + '/public'));
-
-var impObject = {
-    'jwtSecret': 'mockcontactcenter'
-};
-
-app.set('jwtSecret', impObject.jwtSecret);
 
 // configure app to use bodyParser()
 // this will let us get the data from a POST
@@ -43,34 +34,6 @@ require("./ApiRoutes")(express, app);
 
 router.get('/', function (req, res) {
     res.json({ message: 'welcome to our api!' });
-});
-
-router.get('/users', function (req, res) {
-    //console.log(userData.user1);
-    res.json(userData.user1);
-});
-
-router.post('/authenticate', function (req, res) {
-
-    // find the user
-
-    if (userData.user1.name != req.body.name) {
-        res.json({ success: false, message: 'Authentication failed. User not found.' });
-    }
-    if (userData.user1.password != req.body.password) {
-        res.json({ success: false, message: 'Authentication failed. Wrong password.' });
-    }
-    var accessToken = jwt.sign(
-        {name:req.body.name}, 
-        impObject.jwtSecret, 
-        {expiresIn: '1h' }
-    );
-    res.json({
-          success: true,
-          message: 'Enjoy your token!',
-          token: accessToken
-        });
-
 });
 
 app.use('/api', router);
