+++ test/test.js
@@ -22,13 +22,12 @@ describe('Root API', () => {
 describe('authenticate API', () => {
     it('authenticate api testing', function (done) {
         chai.request(server)
-            .post('/api/authenticate')
+            .post('/api/tokenAuthentication')
             .send({ 'name': 'dhrisya', 'password': 'password' })
             .end(function (err, res) {
                 res.should.have.status(200);
                 res.body.should.have.property('message').eql('Enjoy your token!');
                 res.body.should.have.property('token');
-                //console.log(res.body.should.have.property('message'));
                 done();
             });
     });
@@ -38,10 +37,10 @@ describe('authenticate API', () => {
 describe('SendSuggestion API', () => {
     it('Recommendation api testing', function (done) {
         chai.request(server)
-            .post('/api/sendSuggestion')
+            .post('/api/sendRecommendation')
+            .send({ 'AgentId': '46', 'message': 'password' })
             .end(function (err, res) {
                 res.should.have.status(200);
-                res.body.should.have.property('message').eql('Success. Message Sent');
                 done();
             });
     });
